server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.4
